Change arguments of remix.weights
Add arguments return.loglik and w.type_1. The former is a logical that requests object loglik be returned. The latter are the unnormalised type 1 weights. By supplying these both after runing once to calculate the type 1 weights only, we can speed up calculation of the type 2 weights because of shared intermediaries. This means we can get timings for both algorithms and not waste any computation.