Fix a bug in mopp.weights and other changes
There was a bug in mopp.weights affecting weights (all types) when Laplace sampling was in use (all types). This was that the likelihoods for the Laplace samples were being included in the product of likelihoods representing the unnormalised full posterior density, which is incorrect. Changed to just using the samples from the local posteriors only.

Also made mopp.weights more memory efficient when using Laplace samples.